decides not to delegate to essential value
After some careful thought, I've decided Kindness should _not_ delegate
(via #method_missing) all undefined messages to the underlying essence.

There were a few method names which Kindness needs for its own use, but
which could very well be useful for an essential object, such as #valid?
and #messages.  Accordlingly, I've decided here that it's better to just
require the user to extract (via #essence or #value) the essential
object before interacting with it.

The anticipated use case for which Kindness was created (type checking
for pattern-matching functions) should be quite compatible with this
decision; arguments can be type-checked and then extracted before being
handed off as arguments to the target function definition.