renames abstractions
Instead of a class called 'Rubric', naming as a class called 'Signature'
since that's basically what it is.

Instead of a DSL method 'rubric', naming as DSL method 'insist' since
that's basically what we're doing by creating a rule.  Another name
which I considered was 'stipulate'.  I may alias that name. Both of
those are verbs, which I think reads better than using the noun 'rubric'.