Removes Kindness::Signature
The purpose and usecase for signatures was a lot more broad and a lot
less clear than the core, monadic goals of Identity and Criterion.s

The idea of functionality which validates method signatures is nifty,
but it's really biting off a lot more in terms of scope.  Without a
clear usecase towards which to work, I find that scope too nebulous to
define meaninful goals, so I'm setting that scope aside in favor of a
very light and lean core.